Implemented Chunked collisons, gravity...

+++ include/simulation/containers/particle.h
@@ -5,10 +5,11 @@
  */
 
 #include <stdbool.h>
+#include <stdint.h>
 
 typedef struct {
     float pos[3];
     float vel[3];
-    float col[3];
+    uint8_t col[3];
     float radius;
 } Particle;
